[BE] refactor(#268): 피드백 반영
* fcmToken -> fcm 변경
* fcm 도메인에서 listener 관리
* event 객체는 각각의 도메인에서 관리
