+++ backend/src/main/java/com/example/backend/domain/define/fcm/listener/MemberEventListener.java
@@ -1,10 +1,7 @@
-package com.example.backend.domain.define.study.member.listener;
+package com.example.backend.domain.define.fcm.listener;
 
-import com.example.backend.common.exception.ExceptionMessage;
-import com.example.backend.common.exception.event.EventException;
-import com.example.backend.domain.define.fcmToken.FcmToken;
-import com.example.backend.domain.define.fcmToken.repository.FcmTokenRepository;
-import com.example.backend.domain.define.study.member.listener.event.ResignMemberEvent;
+import com.example.backend.domain.define.fcm.FcmToken;
+import com.example.backend.domain.define.study.member.event.ResignMemberEvent;
 import com.example.backend.study.api.event.FcmSingleTokenRequest;
 import com.example.backend.study.api.event.service.FcmService;
 import com.google.firebase.messaging.FirebaseMessagingException;
