+++ backend/src/main/java/com/example/backend/domain/define/fcm/repository/FcmTokenRepository.java
@@ -1,6 +1,6 @@
-package com.example.backend.domain.define.fcmToken.repository;
+package com.example.backend.domain.define.fcm.repository;
 
-import com.example.backend.domain.define.fcmToken.FcmToken;
+import com.example.backend.domain.define.fcm.FcmToken;
 import org.springframework.data.repository.CrudRepository;
 
 public interface FcmTokenRepository extends CrudRepository<FcmToken, Long> {
