+++ backend/src/main/java/com/example/backend/domain/define/study/info/event/ApplyApproveRefuseMemberEvent.java
@@ -1,4 +1,4 @@
-package com.example.backend.domain.define.study.info.listener.event;
+package com.example.backend.domain.define.study.info.event;
 
 
 import lombok.AllArgsConstructor;
@@ -10,12 +10,13 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ApplyMemberEvent {
+public class ApplyApproveRefuseMemberEvent {
 
-    private Long studyLeaderId;  // 스터디 리더 Id
+    private boolean approve;   // 승인 여부
+
+    private Long applyUserId;  // 가입신청자 Id
 
     private String studyTopic;  // 스터디 제목
 
     private String name;      // 가입신청자 이름
-
 }
